+++ src/utils.go
@@ -14,5 +14,5 @@ func dist2D(p0 Coord2D, p1 Coord2D) float32 {
 }
 
 func Point3DToPoint2D(v Vertex, fv float32) Coord2D {
-    return Coord2D{int16(float32(v.X)/float32(v.Z)*fv), int16(float32(v.Y)/float32(v.Z)*fv)};
+    return Coord2D{int16(v.X/v.Z*fv), int16(v.Y/v.Z*fv)};
 }
